requirements.txt
albumentations==1.3.0
easydict
tensorflow==2.14.1
timm
omegaconf
efficientnet_pytorch
kornia
pandas